move test utilities to conda.testing
