+++ tests/test_install.py
@@ -21,7 +21,7 @@ import sys
 import tempfile
 import unittest
 
-from .helpers import mock
+from conda.testing.helpers import mock
 
 patch = mock.patch if mock else None
