+++ tests/test_plan.py
@@ -22,9 +22,9 @@ from conda.models.records import PackageRecord
 from conda.models.match_spec import MatchSpec
 from conda.plan import display_actions, add_unlink, add_defaults_to_specs, _update_old_plan as update_old_plan
 from conda.exports import execute_plan
-from .decorators import skip_if_no_mock
 from .gateways.disk.test_permissions import tempdir
-from .helpers import captured, get_index_r_1, mock, tempdir
+from conda.testing.helpers import captured, get_index_r_1, mock, tempdir
+from conda.testing.decorators import skip_if_no_mock
 
 index, r, = get_index_r_1()
 index = index.copy()  # create a shallow copy so this module can mutate state
